% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{clearNames}
\alias{clearNames}
\title{Retrieve Specie Name from Full Taxonomy}
\usage{
clearNames(specie_names)
}
\arguments{
\item{otu_table}{an OTU table having the full taxonomy string in the columns}
}
\value{
an OTU table having the specie name and its identificative number as column names
}
\description{
The function retrieves the specie name and eventually its identificative number.
It substitutes `Specie incertae sedis` with `sp.`
}
