% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge a list of metadata into a new dataframe. It also mutate Study and Invididual columns as factor}
\usage{
merge_metadata(Metadata_list)
}
\arguments{
\item{Metadata_list}{a list of metadata with unique individuals}
}
\value{
A dataframe with all the metadata in the list
}
\description{
Merge a list of metadata into a new dataframe. It also mutate Study and Invididual columns as factor
}
