% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{runSIAMCAT}
\alias{runSIAMCAT}
\title{Run SIAMcat Linar Mixed Model}
\usage{
runSIAMCAT(phyloseq, metadata, study_coveriate = TRUE)
}
\arguments{
\item{phyloseq}{Phyloseq object}

\item{metadata}{Dataframe with the timepoints before and after and, eventually, study variable}

\item{study_covariate}{include the study variable as a random effect in the linear model}
}
\value{
A siamcat object with the associations stored in it
}
\description{
From the phyloseq and metadata inputs, runs SIAMcat's lmm through the before and after timepoints in the metadata
}
