% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{savePlots}
\alias{savePlots}
\title{Save a list of plots to a pdf}
\usage{
savePlots(plots, filename, width, height)
}
\arguments{
\item{plots}{A list of plots}

\item{filename}{String with the complete path for where do you want the file to be saved}

\item{width}{plot width}

\item{height}{plot height}
}
\value{
A pdf file with one plot for each page
}
\description{
Save a list of plots to a pdf
}
